+++ test/test_integration/test_main.cpp
@@ -22,9 +22,10 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 #include <Arduino.h>
+#include <EEPROM.h>
 #include <unity.h>
 #include <Shell.h>
-#include <TesterStream.h>
+#include "../TesterStream.h"
 
 TesterStream tester;
 TesterStream tester2;
@@ -43,7 +44,7 @@ handler(VER, "Displays firmware version.")
 handler(TEST, "Used for unit testing. <month>")
 {
     int16_t month;
-    int8_t res = request.readInt(&month, 1, 12);
+    int8_t res = request.readInt16(&month, 1, 12);
     if (res <= 0)
         return SHELL_RESPONSE_ERR_BAD_ARGUMENT;
     char *str = request.peek();
@@ -58,7 +59,7 @@ handler(TEST, "Used for unit testing. <month>")
 handler(A, "Short command no dot")
 {
     int16_t code;
-    if (!request.readInt(&code))
+    if (!request.readInt16(&code))
         return 0;
     return code;
 }
